adding new function and modify script

+++ gprs/gprs.py
@@ -90,21 +90,31 @@ class GPRS( object ):
     def generate_plink_bfiles(self, snplist_name, output_name, symbol='.' ):
         for nb in range( 1, 23 ):
             chrnb = "chr{}".format( nb )
-            # get snps lists from directory
             for snps in os.listdir( self.snplists_dir() ):
-                if "{}_{}.csv".format( chrnb, snplist_name ) in snps:
-                    for i in os.listdir( self.ref ):
-                        if i.endswith('.vcf.gz') and chrnb != "chrX" and chrnb != "chrY" and chrnb != "chrMT" and "{}{}".format(chrnb, symbol) in i:
-                            # exclude chr X, Y and MT, and setting a filter to make sure all input are consistent
-                            # ex: chr1 snps-list and chr1 vcf read at the same time
-                            os.system("plink --vcf {}/{} --extract {}/{} --make-bed --out {}/{}_{}".format( self.ref, i,
+                if "chr"  and "{}_{}.csv".format( chrnb, snplist_name ) in snps:
+                        for i in os.listdir( self.ref ):
+                            if i.endswith('.vcf.gz') and chrnb != "chrX" and chrnb != "chrY" and chrnb != "chrMT" and "{}{}".format(chrnb, symbol) in i:
+                                    # exclude chr X, Y and MT, and setting a filter to make sure all input are consistent
+                                    # ex: chr1 snps-list and chr1 vcf read at the same time
+                                os.system("plink --vcf {}/{} --extract {}/{} --make-bed --out {}/{}_{}".format( self.ref, i,
+                                                                                                              self.snplists_dir(),
+                                                                                                              snps,
+                                                                                                              self.plink_bfiles_dir(),
+                                                                                                              chrnb,
+                                                                                                              output_name ) )
+                            print( "with chr: {}_{} is finished!".format( chrnb, output_name ) )
+                elif "chr" not in snps and "{}.csv".format( snplist_name ) in snps:
+                        for i in os.listdir( self.ref ):
+                            if i.endswith('.vcf.gz' ) and chrnb != "chrX" and chrnb != "chrY" and chrnb != "chrMT" and "{}{}".format(chrnb, symbol ) in i:
+                               os.system("plink --vcf {}/{} --extract {}/{} --make-bed --out {}/{}_{}".format( self.ref, i,
                                                                                                       self.snplists_dir(),
                                                                                                       snps,
                                                                                                       self.plink_bfiles_dir(),
                                                                                                       chrnb,
                                                                                                       output_name ) )
-                            print( "{}_{} is finished!".format( chrnb, output_name ) )
+                            print( "no chr: {}_{} is finished!".format( chrnb, output_name ) )
         print( "all jobs completed!" )
+
 
     def clump(self, qc_file_name, plink_bfile_name, output_name, clump_kb, clump_p1, clump_p2, clump_r2='0.1', clump_field='Pvalue',clump_snp_field='SNPID'):
         for nb in range( 1, 23 ):
@@ -115,20 +125,33 @@ class GPRS( object ):
                 if "{}_{}.QC.csv".format( chrnb, qc_file_name ) in qc_files:
                     for plink_files in os.listdir( self.plink_bfiles_dir() ):
                         if "{}_{}.bed".format( chrnb, plink_bfile_name ) in plink_files:
-                            plinkinput = plink_files.split( "." )[0]
-                            # use plink to do the clump
-                            # NOTE: the summary statistics file has to separate by space otherwise plink can not read it
-                            os.system( "plink --bfile {}/{} --clump {}/{} --clump-p1 {} --clump-p2 {} --clump-r2 {} --clump-kb {} --clump-field {} --clump-snp-field {} --out {}/{}_{}".format(
-                                    self.plink_bfiles_dir(), plinkinput,
-                                    self.qc_dir(), qc_files,
-                                    clump_p1, clump_p2, clump_r2,
-                                    clump_kb, clump_field, clump_snp_field,
-                                    self.plink_clump_dir(), chrnb, output_name ) )
+                                        plinkinput = plink_files.split( "." )[0]
+                                        # use plink to do the clump
+                                        # NOTE: the summary statistics file has to separate by space otherwise plink can not read it
+                                        os.system( "plink --bfile {}/{} --clump {}/{} --clump-p1 {} --clump-p2 {} --clump-r2 {} --clump-kb {} --clump-field {} --clump-snp-field {} --out {}/{}_{}".format(
+                                                self.plink_bfiles_dir(), plinkinput,
+                                                self.qc_dir(), qc_files,
+                                                clump_p1, clump_p2, clump_r2,
+                                                clump_kb, clump_field, clump_snp_field,
+                                                self.plink_clump_dir(), chrnb, output_name ) )
+                elif "{}.QC.csv".format( qc_file_name ) in qc_files:
+                    for plink_files in os.listdir( self.plink_bfiles_dir() ):
+                        if "{}_{}.bed".format( chrnb, plink_bfile_name ) in plink_files:
+                                        plinkinput = plink_files.split( "." )[0]
+                                        # use plink to do the clump
+                                        # NOTE: the summary statistics file has to separate by space otherwise plink can not read it
+                                        os.system( "plink --bfile {}/{} --clump {}/{} --clump-p1 {} --clump-p2 {} --clump-r2 {} --clump-kb {} --clump-field {} --clump-snp-field {} --out {}/{}_{}".format(
+                                                self.plink_bfiles_dir(), plinkinput,
+                                                self.qc_dir(), qc_files,
+                                                clump_p1, clump_p2, clump_r2,
+                                                clump_kb, clump_field, clump_snp_field,
+                                                self.plink_clump_dir(), chrnb, output_name ) )
         print( "Finished Plink clump!" )
 
     def select_clump_snps(self, qc_file_name, clump_file_name, output_name):
         for nb in range( 1, 23 ):
             chrnb = "chr{}".format( nb )
+            #extract clumped snplist
             try:
                 df = pd.read_csv( "{}/{}_{}.clumped".format( self.plink_clump_dir(), chrnb, output_name ), delim_whitespace=True )
                 new_df = df.loc[(df['CHR'] == nb)]
@@ -136,13 +159,24 @@ class GPRS( object ):
                 new_df_2.to_csv("{}/{}_{}_clumped_snplist.csv".format( self.plink_clump_dir(), chrnb, output_name ), sep=' ', index=False, header=True )
             except IOError:
                 print("{}/{}_{}.clumped not found. skip".format( self.plink_clump_dir(), chrnb, output_name ))
+            #generate snplist by comparing clumped/qced snplists
             try:
                 clump_snp = pd.read_csv("{}/{}_{}_clumped_snplist.csv".format( self.plink_clump_dir(), chrnb, clump_file_name ), sep=' ' )
                 clump_snp.rename( columns={'SNP': 'SNPID'}, inplace=True )
-                qc_snp = pd.read_csv( "{}/{}_{}.QC.csv".format( self.qc_dir(), chrnb, qc_file_name ), sep=' ' )
-                newsnplist = qc_snp[qc_snp["SNPID"].isin( clump_snp["SNPID"] )]
-                newsnplist.to_csv( "{}/{}_{}.qc_clump_snpslist.csv".format( self.qc_clump_snpslist_dir(), chrnb, output_name ), sep=' ', index=False, header=True )
-                print( "{} snpslist has been created".format( chrnb ) )
+                qc_files1 = "{}/{}_{}.QC.csv".format( self.qc_dir(), chrnb, qc_file_name )
+                qc_files2 = "{}/{}.QC.csv".format( self.qc_dir(), qc_file_name )
+                if os.path.exists(qc_files1):
+                    print("start extracting {}".format(qc_files1))
+                    qc_snp = pd.read_csv( "{}/{}_{}.QC.csv".format( self.qc_dir(), chrnb, qc_file_name ), sep=' ' )
+                    newsnplist = qc_snp[qc_snp["SNPID"].isin( clump_snp["SNPID"] )]
+                    newsnplist.to_csv( "{}/{}_{}.qc_clump_snpslist.csv".format( self.qc_clump_snpslist_dir(), chrnb, output_name ), sep=' ', index=False, header=True )
+                    print( "{} snpslist has been created".format( chrnb ) )
+                elif os.path.exists(qc_files2):
+                    print( "start extracting {}".format( qc_files2 ) )
+                    qc_snp = pd.read_csv( "{}/{}.QC.csv".format( self.qc_dir(), qc_file_name ), sep=' ' )
+                    newsnplist = qc_snp[qc_snp["SNPID"].isin( clump_snp["SNPID"] )]
+                    newsnplist.to_csv("{}/{}_{}.qc_clump_snpslist.csv".format( self.qc_clump_snpslist_dir(), chrnb, output_name ), sep=' ', index=False, header=True )
+                    print( "{} snpslist has been created".format( chrnb ) )
             except IOError:
                 print("{}/{}_{}_clumped_snplist.csv not found. skip".format( self.plink_clump_dir(), chrnb, clump_file_name ))
         print( "All jobs are completed" )
@@ -164,3 +198,47 @@ class GPRS( object ):
                     except IOError:
                         print( "{}/{}_{}.qc_clump_snpslist.csv not found. skip".format( self.qc_clump_snpslist_dir(), chrnb, qc_file_name ) )
         print( "ALL work are complete!" )
+
+    def combine_prs(self,pop):
+        ##created by - Soyoung Jeon
+        ###calculate PRS from .sscore files (plink2 output)
+        # needs PREFIX_chrX.sscore files as inputs
+        # outputs PREFIX.sscore with same prefix
+        # usage: python [prefix before '_chrX.sscore' separated by comma]
+
+        POP = pop.split( ',' )
+
+        chr = [i for i in range( 1, 23 )]
+        for p in POP:
+            nm = {}  # NMISS_ALLELE_CT
+            sumsc = {}  # score as SUM of dosage*score
+            avg = {}
+            # make dictionary for each sample
+            f = open( "{}/chr10_{}.sscore".format(self.prs_dir(), p) )
+            f.readline()
+            for lines in f:
+                lines = lines.split()
+                nm[lines[0]] = []
+                sumsc[lines[0]] = []
+                avg[lines[0]] = []
+            # sum PRS per chromosome
+            for i in chr:
+                i = str( i )
+                try:
+                    f = open( "{}/chr{}_{}.sscore".format(self.prs_dir(), i, p)  )
+                    f.readline()
+                    for lines in f:
+                        lines = lines.split()
+                        id = lines[0]
+                        nm[id].append( float( lines[1] ) )
+                        sumsc[id].append( float( lines[1] ) * float( lines[3] ) )
+                        avg[id].append( float( lines[3] ) )
+                except IOError:
+                    print("'File Not Found:'{}/chr{}_{}.sscore'".format(self.prs_dir(),i,p))
+            # sum up scores per chromosome and output
+            o = open( p + '.sscore', 'w' )
+            o.write( 'id\tALLELE_CT\tSCORE_AVG\tSCORE_SUM\n' )
+            for i in nm.keys():
+                o.write( i + '\t' + str( sum( nm[i] ) ) + '\t' + str( sum( avg[i] ) / 22 ) + '\t' + str(
+                    sum( sumsc[i] ) ) + '\n' )
+            o.close()
